% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessdiff.R
\name{compute_accessdiff}
\alias{compute_accessdiff}
\title{Compute mean difference in accessibility}
\usage{
compute_accessdiff(project_tiff, base_tiff, weight_tiff = NULL)
}
\arguments{
\item{weight_tiff}{A Conveyal raster layer containing origin-side
opportunities to use for weighting.}

\item{difference_tiff}{A Conveyal raster layer containing accessibility
statistics for a transit system with an added project, differenced from a
base layer.}
}
\description{
This function computes the average accessibility gain (positive changes
only), optionally weighted by another statistic.
}
\examples{
clifton <- read_tiff(system.file("extdata", "clifton_am_jobs.tiff", package = "tvmeasures"))
base <- read_tiff(system.file("extdata", "base_am_jobs.tiff", package = "tvmeasures"))
compute_accessdiff(clifton, base)


}
