% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitive_uses.R
\name{cultenv}
\alias{cultenv}
\title{Compute cultural and environmental resources measure}
\usage{
cultenv(project, sensitive_layer, buffer, epsg = NULL)
}
\arguments{
\item{project}{An object of class `sf` identifing a project alignment}

\item{sensitive_layer}{An object of class `sf` defining sensitive land uses.}

\item{buffer}{Distance considered an impact on point and linear uses, in the
units of `epsg`}

\item{epsg}{The EPSG code of the reprojected layer.}
}
\value{
The percent of the project's alignment lying in sensitive areas
}
\description{
Compute cultural and environmental resources measure
}
\details{
The cultural and environmental resources measure is the share of a project's
 alignment lying within sensitive land uses. These land uses are defined by
 areal, linear, or point geometries; in the case of points and lines, a buffer
 is applied to the geometries.

The buffer radius calculation requires that the project and sensitivity layer be
 in the same projected coordinate system.
}
