% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessdiff.R
\name{get_tweights}
\alias{get_tweights}
\title{Get weights from a tiff}
\usage{
get_tweights(weight_tiff = NULL, values)
}
\arguments{
\item{weight_tiff}{A Conveyal raster layer containing origin-side
opportunities to use for weighting.}

\item{values}{Raster cell values}
}
\description{
Get weights from a tiff
}
