% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessdiff.R
\name{gini_access}
\alias{gini_access}
\title{Compute the Gini coefficient for a raster}
\usage{
gini_access(tiff, weight_tiff = NULL, ...)
}
\arguments{
\item{tiff}{A Conveyal tiff with accessibility scores.}

\item{weight_tiff}{A Conveyal raster layer containing origin-side
opportunities to use for weighting.}

\item{...}{Arguments passed on to \code{ineq::ineq}
\describe{
  \item{x}{a vector containing at least non-negative elements}
  \item{parameter}{parameter of the inequality measure (if set to \code{NULL}
  the default parameter of the respective measure is used)}
  \item{type}{character string giving the measure used to compute inequality.
  must be one of the strings in the default argument (the first character
  is sufficient). defaults to "Gini".}
  \item{na.rm}{logical. Should missing values (\code{NA}s) be removed
  prior to computations? If set to \code{FALSE} the computations yield
  \code{NA}.}
}}
}
\description{
Compute the Gini coefficient for a raster
}
\details{
The [Gini coefficient](https://en.wikipedia.org/wiki/Gini_coefficient)
is a measure of inequality usually applied to income. In the case of transit
accessibility, Gini values closer to 1 imply that some areas have
high-quality access while most have none. This serves as a wrapper to \link[ineq][ineq]
}
