% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_mgmt.R
\name{landuse_comp}
\alias{landuse_comp}
\title{Compute the Land Use Compatibility Measure}
\usage{
landuse_comp(pop, jobs)
}
\arguments{
\item{pop}{The population within 1/2 mile of proposed transit project
stops or stations with new access to transit.}

\item{jobs}{The jobs within 1/2 mile of proposed transit project
stops or stations with new access to transit.}
}
\value{
A numeric value of the balance between jobs and population within
  1/2 mile of project stations.
}
\description{
Compute the Land Use Compatibility Measure
}
