% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_raster.R
\name{leaflet_raster}
\alias{leaflet_raster}
\title{Plot a raster image as a leaflet map}
\usage{
leaflet_raster(raster, project, diff = TRUE, cuts = c(10, 1000),
  base_rail = marta_sf)
}
\arguments{
\item{raster}{The raster object, projected in WGS84}

\item{project}{A sf object with the project definition}

\item{diff}{If TRUE (default), use a diverging color scheme.}

\item{cuts}{The breaks to use (negative and positive) in showing the colors.
Must be a vector of two increasing numbers.}

\item{base_rail}{An sf table with existing rail or fixed-guideway projects to
show in the background. Defaults to included MARTA lines}
}
\description{
Plot a raster image as a leaflet map
}
