% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marta_sf}
\alias{marta_sf}
\title{MARTA alignments}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 5 rows and 8 columns.}
\usage{
marta_sf
}
\description{
A dataset containing the alignments of MARTA rail systems in Atlanta.
}
\keyword{datasets}
