% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_project.R
\name{read_project}
\alias{read_project}
\title{Read a project geojson file}
\usage{
read_project(path)
}
\arguments{
\item{path}{Path to a project geojson file.}
}
\value{
An object of class \link[sf]{sf} including stop locations and route
  alignment for a project
}
\description{
`read_project` brings a project defined in a geojson file into the workspace.
}
\examples{
proj_file <- system.file("extdata", "clifton.geojson", package = "tvmeasures")
clifton <- read_project(proj_file)
plot(clifton)

}
