% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tiff.R
\name{read_tiff}
\alias{read_tiff}
\title{Read a Conveyal Regional Analysis .tiff file}
\usage{
read_tiff(path)
}
\arguments{
\item{path}{Path to file}
}
\description{
Read a Conveyal Regional Analysis .tiff file
}
\examples{
base_file <- system.file("extdata", "base_am_jobs.tiff", package = "tvmeasures")
base_am_jobs <- read_tiff(base_file)

}
