% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_evaluation_app}
\alias{run_evaluation_app}
\title{Run the shiny app to process the evaluation measures}
\usage{
run_evaluation_app()
}
\description{
Run the shiny app to process the evaluation measures
}
