% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_mgmt.R
\name{social_equity}
\alias{social_equity}
\title{Compute the Social Equity Measure}
\usage{
social_equity(pct_minority = 0, pct_poverty = 0)
}
\arguments{
\item{pct_minority}{The percent of the accessed population that is minority}

\item{pct_poverty}{The percent of the accessd population that is under the poverty line}
}
\description{
Compute the Social Equity Measure
}
