% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessdiff.R
\name{sum_access}
\alias{sum_access}
\title{Sum an accessibility matrix}
\usage{
sum_access(tiff, weight_tiff = NULL)
}
\arguments{
\item{tiff}{A Conveyal tiff with accessibility scores.}

\item{weight_tiff}{A Conveyal raster layer containing origin-side
opportunities to use for weighting.}
}
\description{
Sum an accessibility matrix
}
