% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessdiff.R
\name{table_builder}
\alias{table_builder}
\title{Function to make a Conveyal Results table}
\usage{
table_builder(results, base_tiff, weight_tiff, weight_prefix = "",
  probs = c(0.7, 0.9))
}
\arguments{
\item{results}{A conveyal regional analysis raster object.}

\item{base_tiff}{The base scenario raster object.}

\item{weight_tiff}{The raster object with population weights.}

\item{weight_prefix}{A character string to designate the type of weighting.}

\item{probs}{Vector of percentiles to use to compute accessibility.}
}
\value{
A data_frame with the sum, total percent change from base, and
  percentile accessibilities for a given weighting regimen.
}
\description{
Function to make a Conveyal Results table
}
